STAR-543 Port remaining guardrails and sync with cndb (#180)
(cherry picked from commit a83e0e87c5521ec24e76de57db03dc0ed093c73f)
(cherry picked from commit b90b43306e89172ae7882cb041157abf7e30740e)
(cherry picked from commit 36600dfeadaad6affe2b264e3e8f07853b7c94df)
(cherry picked from commit abc39d60b9b82178d86cfc86c880ea7978835fe0)
(cherry picked from commit 7a263dcf3f827ea68f10de04d964da0f84f851ab)
(cherry picked from commit 159a1fa87ed7b6e4696b91bfd60eb58492679e5e)
(cherry picked from commit 503767436b374646362c1adfb3abd3ef181d8484)
(cherry picked from commit ab880d19b9c4dedc551191a6e944d83014fa56a8)

STAR-543 Fix rebase compile errors and adapt new guardrails for OS 5.0 implementation;
Ported CC implementation of guardrails to OS - loggedBatchEnabled, scannedTombstones, batchSize, and unloggedBatchAcrossPartitions;
GuardedCounter has been ported from CC and needed a temporary fix to support the situation in ReadCommand where a ClientState is not readily available;
From CC, the GuardrailLoggedBatchTest has been added to o.a.c.db.guardrails package

STAR-543 Fix ClientWarningsTest.testTombstoneWarning

STAR-543 Slightly better fix to allow GuardedCounter.add to trigger a failure and raise a GuardrailViolatedException.

STAR-543 Ignore TombstoneCountWarningTest because scannedTombstones guardrail replaces tombstone warnings from CASSANDRA-16896

STAR-543 Add cassandra.enable_guardrails_for_anonymous_user property to allow CC to enable guardrails when authentication is not used.
In GuardrailCollectionSizeOnSSTableWriterTest.testMapSize ensure inserted values are unique to bypass NoSpamLogger shouldLog checks.

STAR-543 Fix GuardrailWriteConsistencyLevelsTest failures
Remove guardrail checks from ConsistencyLevel that duplicate guardrail checks already in Statement classes;
Change expectations in GuardrailWriteConsistencyLevelsTest.testLWTQuery when null serial CL is used to reflect the ConsistencyLevel.defaultSerialConsistency method results.

+++ src/java/org/apache/cassandra/metrics/Sampler.java
@@ -106,7 +106,7 @@ public abstract class Sampler<T>
 
     private long endTimeNanos = -1;
 
-    public void addSample(final T item, final int value)
+    public void addSample(final T item, final long value)
     {
         if (isEnabled())
             samplerExecutor.submit(() -> insert(item, value));
